# Spring III/docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"pf-be-java-hisp-w18-g1","description":"pf-be-java-hisp-w18-g1","contact":{"name":"ext_caarroyo","email":"ext_caarroyo@mercadolibre.com"},"version":"last"},"servers":[{"url":"http://localhost:8080","description":"Scope local"}],"paths":{"/api/v1/fresh-products/orders/{orderId}":{"put":{"tags":["purchase-order-controller"],"operationId":"updatePurchaseOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderCreationResponseDTO"}}}}}}},"/api/v1/fresh-products/inboundorder":{"put":{"tags":["inbound-order-controller"],"operationId":"updateInboundOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatchStockResponseDTO"}}}}}},"post":{"tags":["inbound-order-controller"],"operationId":"addInboundOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatchStockResponseDTO"}}}}}}},"/login":{"post":{"tags":["session-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationUserRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApplicationUserResponseDTO"}}}}}}},"/api/v1/fresh-products/orders":{"post":{"tags":["purchase-order-controller"],"operationId":"addPurchaseOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderCreationResponseDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/buyer":{"get":{"tags":["roles-test-controller"],"operationId":"authBuyer","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/buyer2":{"get":{"tags":["roles-test-controller"],"operationId":"authBuyer2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/agent":{"get":{"tags":["roles-test-controller"],"operationId":"authAgent","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/agent2":{"get":{"tags":["roles-test-controller"],"operationId":"authAgent2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products/{idProduct}/warehouse/list":{"get":{"tags":["product-controller"],"operationId":"getAllProduct","parameters":[{"name":"idProduct","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/fresh-products/{idProduct}/batch/list":{"get":{"tags":["product-controller"],"operationId":"getAllProductOrderBy","parameters":[{"name":"idProduct","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"order","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/fresh-products/orders/{order}":{"get":{"tags":["purchase-order-controller"],"operationId":"getProductsByPurchaseOrder","parameters":[{"name":"order","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}}}}}},"/api/v1/fresh-products/list":{"get":{"tags":["product-controller"],"operationId":"getAllByCategory","parameters":[{"name":"category","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}}}}}},"/api/v1/fresh-products/batch/list/due-date/{cantDays}":{"get":{"tags":["batch-controller"],"operationId":"getRecentPostsFromFollowed","parameters":[{"name":"cantDays","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string"}},{"name":"order","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatchStockDTO"}}}}}}}},"components":{"schemas":{"ItemDTO":{"type":"object","properties":{"quantity":{"type":"integer","format":"int32"},"product_id":{"type":"integer","format":"int64"}}},"OrderStatusDTO":{"type":"object","properties":{"status_code":{"type":"string"}}},"PurchaseOrderDTO":{"type":"object","properties":{"date":{"type":"string","format":"date"},"buyer_id":{"type":"integer","format":"int64"},"order_status":{"$ref":"#/components/schemas/OrderStatusDTO"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ItemDTO"}}}},"PurchaseOrderRequest":{"type":"object","properties":{"purchase_order":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}},"PurchaseOrderCreationResponseDTO":{"type":"object","properties":{"order_number":{"type":"integer","format":"int64"},"total_price":{"type":"number","format":"double"}}},"BatchDTO":{"type":"object","properties":{"batch_number":{"type":"integer","format":"int64"},"product_id":{"type":"integer","format":"int64"},"current_temperature":{"type":"number","format":"double"},"minimum_temperature":{"type":"number","format":"double"},"initial_quantity":{"type":"integer","format":"int32"},"current_quantity":{"type":"integer","format":"int32"},"manufacturing_date":{"type":"string","format":"date"},"manufacturing_time":{"type":"string","format":"date-time"},"due_date":{"type":"string","format":"date"}}},"InboundOrderDTO":{"type":"object","properties":{"section":{"$ref":"#/components/schemas/SectionDTO"},"order_number":{"type":"integer","format":"int64"},"order_date":{"type":"string","format":"date"},"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/BatchDTO"}}}},"InboundOrderRequest":{"type":"object","properties":{"inbound_order":{"$ref":"#/components/schemas/InboundOrderDTO"}}},"SectionDTO":{"type":"object","properties":{"section_code":{"type":"integer","format":"int32"},"warehouse_code":{"type":"integer","format":"int64"}}},"BatchStockResponseDTO":{"type":"object","properties":{"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/BatchDTO"}}}},"ApplicationUserRequestDTO":{"type":"object","properties":{"password":{"type":"string"},"user_name":{"type":"string"}}},"ApplicationUserResponseDTO":{"type":"object","properties":{"token":{"type":"string"},"user_id":{"type":"integer","format":"int64"},"user_name":{"type":"string"}}},"ProductDTO":{"type":"object","properties":{"description":{"type":"string"},"section_code":{"type":"string","enum":["FRESH","COOL","FROZEN"]},"price":{"type":"number","format":"double"}}},"BatchResponseDTO":{"type":"object","properties":{"batch_number":{"type":"integer","format":"int64"},"product_id":{"type":"integer","format":"int64"},"product_type_id":{"type":"integer","format":"int32"},"due_date":{"type":"string","format":"date"},"current_quantity":{"type":"integer","format":"int32"}}},"BatchStockDTO":{"type":"object","properties":{"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/BatchResponseDTO"}}}}}}}