# Spring III/src/main/java/com/mercadolibre/pf_be_java_hisp_w18_g1/dtos/ApplicationUserRequestDTO.java
package com.mercadolibre.pf_be_java_hisp_w18_g1.dtos;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ApplicationUserRequestDTO {
    @JsonProperty("user_name")
    private String userName;
    private String password;
}