# Spring III/src/main/java/com/mercadolibre/pf_be_java_hisp_w18_g1/dtos/PurchaseOrderDTO.java
package com.mercadolibre.pf_be_java_hisp_w18_g1.dtos;

import com.fasterxml.jackson.annotation.JsonFormat;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.format.annotation.DateTimeFormat;

import java.time.LocalDate;
import java.util.List;

@NoArgsConstructor
@AllArgsConstructor
@Data
public class PurchaseOrderDTO {
    @JsonProperty("date")
    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
    @DateTimeFormat(pattern = "dd-MM-yyyy")
    private LocalDate orderDate;
    @JsonProperty("buyer_id")
    private Long buyerId;
    @JsonProperty("order_status")
    private OrderStatusDTO orderStatus;
    @JsonProperty("products")
    private List<ItemDTO> items;
}